+++ software/Core/Inc/pid.h
@@ -21,8 +21,8 @@ static inline void pid_init(void)
  * @param servo_x: X-axis servo control
  * @param servo_y: Y-axis servo control
  */
-void compute_control(int16_t x,
-                     int16_t y,
+void compute_control(int8_t x,
+                     int8_t y,
                      uint16_t *servo_x,
                      uint16_t *servo_y,
                      uint8_t k_p,
